F #1550: Update host network BW units and remove debug files  (#3570)
* F #1550: update host network BW units

* F #1550: remove debug files

+++ src/schedm_mad/remotes/one_drs/lib/models/host.py
@@ -720,7 +720,7 @@ class Host:
                     "type": "Element",
                 },
             )
-            netrx_bw: Optional[int] = field(
+            netrx_bw: Optional[Decimal] = field(
                 default=None,
                 metadata={
                     "name": "NETRX_BW",
@@ -748,7 +748,7 @@ class Host:
                     "type": "Element",
                 },
             )
-            nettx_bw: Optional[int] = field(
+            nettx_bw: Optional[Decimal] = field(
                 default=None,
                 metadata={
                     "name": "NETTX_BW",
